requirements.txt
flask
transformers
torch
pinecone-client
sentence-transformers